test(NES-1727): fold Conductor.perCardChat.spec into Conductor.spec
PR review: the Component.<feature>.spec.tsx split is unique to the chat
workstream — only 3 files repo-wide use it, all from this effort; every
other FE component co-locates in a single Component.spec.tsx. This PR
already rewrites the describe block, so it's a low-cost spot to stop the
pattern from spreading.

Fold the per-card chat tests back into Conductor.spec.tsx as a nested
describe. The PinnedChatBar/ChatOverlay stubs move to file level (harmless
to the non-chat tests — they render without the apologistChat flag, so the
bar never mounts). The folded tests use variant 'customize', so Conductor's
journeyView effect early-returns and never touches the module-level fetch.

Follow-up for the other two (Conductor.apologistChat.spec.tsx +
StepFooter.apologistChat.spec.tsx): NES-1733.
